refactor: /api/studies//contents?progressId= 수정
